test: Move backend-agnostic conformance tests to AbstractConditionalOperationStorageClientTest
Migrate four tests that previously lived only in GcsStorageClientTest
into AbstractConditionalOperationStorageClientTest so every backend
exercises them:

- listBlobs returns all blobs without duplicates
- writeBlobIfGeneration with stale generation throws
- writeBlobIfGeneration with matching generation overwrites
- writeBlobIfGeneration rejects negative expectedGeneration

The two CAS tests need to read the storage-backend generation, which
StorageClient.Blob does not expose. Add a `protected abstract
getGeneration(blobKey)` to the conformance base; each concrete test
overrides it with the backend-appropriate accessor:

- GcsStorageClientTest: storageEmulator.storage.get(...).generation
- FileSystemStorageClientTest: file.lastModified()
- InMemoryStorageClientTest: storageClient.getGenerationForTesting(key)
- Aead/StreamingAead/Mesos wrapper tests: wrappedStorageClient.getGenerationForTesting

InMemoryStorageClient gains a test-only `getGenerationForTesting`
accessor for the abstract base to delegate to.

Remaining backend-specific tests stay put:
- GcsStorageClientTest: `PATCH with null metadata value deletes the key on GCS`
  (GCS REST/SDK quirk, not a contract of the StorageClient interface).
- InMemoryStorageClientTest: `contents()` (InMemory-specific accessor).
- FileSystemStorageClientTest: `writeBlob writes blob to file in subdirectory`
  (filesystem path resolution check).

Issue: None
