+++ src/test/kotlin/org/wfanet/measurement/common/crypto/tink/AeadStorageClientTest.kt
@@ -47,6 +47,9 @@ class AeadStorageClientTest : AbstractConditionalOperationStorageClientTest<Aead
     storageClient = AeadStorageClient(wrappedStorageClient, aead)
   }
 
+  override suspend fun getGeneration(blobKey: String): Long =
+    wrappedStorageClient.getGenerationForTesting(blobKey)
+
   @Test
   fun `wrapped blob is encrypted`() = runBlocking {
     val blobKey = "kms-blob"
