+++ src/test/kotlin/org/wfanet/measurement/common/crypto/tink/StreamingAeadStorageClientTest.kt
@@ -61,6 +61,9 @@ class StreamingAeadStorageClientTest :
     storageClient = StreamingAeadStorageClient(wrappedStorageClient, streamingAead)
   }
 
+  override suspend fun getGeneration(blobKey: String): Long =
+    wrappedStorageClient.getGenerationForTesting(blobKey)
+
   @Test
   fun `wrapped blob is encrypted`() = runBlocking {
     val blobKey = "kms-blob"
