+++ src/test/kotlin/org/wfanet/measurement/gcloud/gcs/GcsStorageClientTest.kt
@@ -17,12 +17,8 @@ package org.wfanet.measurement.gcloud.gcs
 import com.google.cloud.storage.BlobId
 import com.google.cloud.storage.BlobInfo
 import com.google.common.truth.Truth.assertThat
-import com.google.protobuf.ByteString
 import com.google.protobuf.kotlin.toByteStringUtf8
 import java.time.Instant
-import kotlin.test.assertFailsWith
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
 import org.junit.After
 import org.junit.Before
@@ -31,7 +27,6 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 import org.wfanet.measurement.gcloud.gcs.testing.StorageEmulatorRule
-import org.wfanet.measurement.storage.BlobChangedException
 import org.wfanet.measurement.storage.testing.AbstractBlobMetadataStorageClientTest
 import org.wfanet.measurement.storage.testing.BlobSubject.Companion.assertThat
 
@@ -42,6 +37,9 @@ class GcsStorageClientTest : AbstractBlobMetadataStorageClientTest<GcsStorageCli
     storageEmulator.createBucket(BUCKET)
     storageClient = GcsStorageClient(storageEmulator.storage, BUCKET)
   }
+
+  override suspend fun getGeneration(blobKey: String): Long =
+    storageEmulator.storage.get(BUCKET, blobKey).generation
 
   @After
   fun deleteBucket() {
@@ -66,23 +64,6 @@ class GcsStorageClientTest : AbstractBlobMetadataStorageClientTest<GcsStorageCli
       val actualMetadata = blob.metadata ?: emptyMap()
       assertThat(actualMetadata).containsAtLeastEntriesIn(expectedMetadata)
     }
-  }
-
-  @Test
-  fun `listBlobs returns all blobs without duplicates`() = runBlocking {
-    val totalBlobs = 25
-    val prefix = "test/"
-
-    for (i in 1..totalBlobs) {
-      val key = "${prefix}file_${"%04d".format(i)}.json"
-      storageClient.writeBlob(key, flowOf(ByteString.copyFromUtf8("x")))
-    }
-
-    val results = storageClient.listBlobs(prefix).toList()
-    val keys = results.map { it.blobKey }
-
-    assertThat(keys).hasSize(totalBlobs)
-    assertThat(keys.toSet()).hasSize(totalBlobs)
   }
 
   /**
@@ -110,37 +91,6 @@ class GcsStorageClientTest : AbstractBlobMetadataStorageClientTest<GcsStorageCli
 
     val refetched = checkNotNull(storageClient.getBlob(blobKey))
     assertThat(refetched.metadata).doesNotContainKey("foo")
-  }
-
-  @Test
-  fun `writeBlobIfGeneration with stale generation throws`(): Unit = runBlocking {
-    val blobKey = "stale-gen-cas"
-    val staleBlob = storageClient.writeBlob(blobKey, "v1".toByteStringUtf8())
-    val staleGen = storageEmulator.storage.get(BUCKET, blobKey).generation
-    // Concurrent writer races ahead.
-    storageClient.writeBlob(blobKey, "v1.5".toByteStringUtf8())
-
-    assertFailsWith<BlobChangedException> {
-      storageClient.writeBlobIfGeneration(blobKey, staleGen, flowOf(testBlobContent))
-    }
-  }
-
-  @Test
-  fun `writeBlobIfGeneration with matching generation overwrites`(): Unit = runBlocking {
-    val blobKey = "matching-gen-cas"
-    storageClient.writeBlob(blobKey, "v1".toByteStringUtf8())
-    val currentGen = storageEmulator.storage.get(BUCKET, blobKey).generation
-
-    val second = storageClient.writeBlobIfGeneration(blobKey, currentGen, flowOf(testBlobContent))
-
-    assertThat(second).contentEqualTo(testBlobContent)
-  }
-
-  @Test
-  fun `writeBlobIfGeneration rejects negative expectedGeneration`(): Unit = runBlocking {
-    assertFailsWith<IllegalArgumentException> {
-      storageClient.writeBlobIfGeneration("k", expectedGeneration = -1L, flowOf(testBlobContent))
-    }
   }
 
   companion object {
